fix: Resolve compiler warnings in result_sorter and query_cache_test
Address CI/compiler warnings with targeted fixes:

- src/query/result_sorter.cpp: Replace PRIu64 macro with static_cast<unsigned long long>
  for uint64_t formatting to fix CI format warnings. Remove <cinttypes> include as
  it's no longer needed.

- tests/cache/query_cache_test.cpp: Remove unused atomic variable thread1_released_lock
  from ABAProofLockUpgrade test to eliminate compiler warnings.

Both changes resolve warning-level issues identified in CI builds without affecting
functionality.

+++ src/query/result_sorter.cpp
@@ -53,7 +53,7 @@ std::string ResultSorter::GetSortKey(DocId doc_id, const storage::DocumentStore&
           // NOLINTBEGIN(cppcoreguidelines-avoid-c-arrays,modernize-avoid-c-arrays)
           char buf[kNumericBufferSize];
           // NOLINTNEXTLINE(cppcoreguidelines-pro-type-vararg,cppcoreguidelines-pro-bounds-array-to-pointer-decay)
-          snprintf(buf, sizeof(buf), "%020llu", num);
+          snprintf(buf, sizeof(buf), "%020llu", static_cast<unsigned long long>(num));
           // NOLINTEND(cppcoreguidelines-avoid-c-arrays,modernize-avoid-c-arrays)
           return {buf};  // NOLINT(cppcoreguidelines-pro-bounds-array-to-pointer-decay)
         } catch (const std::exception&) {
